+++ tests/Pest.php
@@ -13,6 +13,8 @@
 
 pest()->extend(Itsmurumba\Mpesa\Tests\TestCase::class)->in('Feature', 'Unit');
 
+require_once __DIR__ . '/Unit/Mpesa/_helpers.php';
+
 /*
 |--------------------------------------------------------------------------
 | Expectations
